clients/customize: breakout public page layout
